refactor(snddrv): extracted some functions
- snd_osc_period()
- snd_channel_synthesis()
